+++ front-end/src/api/UserApi.js
@@ -18,6 +18,7 @@ export const insertSignupForm = (data) => {
 
 //user information API - 프로필 조회
 export const selectUserInformation = async (token) => {
+  console.log(token);
   const result = axios
     .get(baseUrl + "profile", {
       headers: {
